refactor: 인증과 사용자 파일 분리
