+++ src/main/java/com/dku/springstudy/dto/auth/request/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package com.dku.springstudy.dto.user.request;
+package com.dku.springstudy.dto.auth.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
