+++ src/main/java/com/dku/springstudy/dto/auth/response/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.dku.springstudy.dto.user.response;
+package com.dku.springstudy.dto.auth.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
