--- src/main/java/com/dku/springstudy/dto/user/UserDto.java
@@ -1,20 +1,0 @@
-package com.dku.springstudy.dto.user;
-
-import com.dku.springstudy.domain.constant.Role;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Getter
-public class UserDto {
-
-    private Long id;
-    private String email;
-    private String password;
-    private String name;
-    private String phoneNumber;
-    private String nickname;
-    private int status;
-    private Role role;
-}
